Introduce a Runtime.IO module for boot etc.
