+++ src/Curiosity/Runtime/Type.hs
@@ -5,26 +5,12 @@ module Curiosity.Runtime.Type
   , rDb
   , rLoggers
   , Threads(..)
-  -- * managing runtimes: boot, shutdown etc. 
-  , boot
-  , boot'
-  , instantiateDb
-  , readDb
-  , readDbSafe
   ) where
 
 import qualified Commence.Multilogging         as ML
-import qualified Commence.Runtime.Errors       as Errs
-import qualified Control.Concurrent.STM        as STM
 import           Control.Lens
 import qualified Curiosity.Core                as Core
-import qualified Curiosity.Data                as Data
 import qualified Curiosity.Parse               as Command
-import qualified Curiosity.Runtime.Error       as RErr
-import qualified Data.Text                     as T
-import qualified Data.Text.Encoding            as TE
-import qualified Data.Text.IO                  as T
-import           System.Directory               ( doesFileExist )
 
 
 --------------------------------------------------------------------------------
@@ -52,106 +38,3 @@ data Threads =
     }
 
 makeLenses ''Runtime
-
-
---------------------------------------------------------------------------------
--- | Boot up a runtime.
-boot
-  :: MonadIO m
-  => Command.Conf -- ^ configuration to boot with.
-  -> Threads
-  -> m (Either Errs.RuntimeErr Runtime)
-boot _rConf _rThreads =
-  liftIO
-      (  try @SomeException
-      .  ML.makeDefaultLoggersWithConf
-      $  _rConf
-      ^. Command.confLogging
-      )
-    >>= \case
-          Left loggerErrs ->
-            putStrLn @Text
-                (  "Cannot instantiate, logger instantiation failed: "
-                <> show loggerErrs
-                )
-              $> Left (Errs.RuntimeException loggerErrs)
-          Right _rLoggers -> do
-            eDb <- instantiateDb _rConf
-            pure $ case eDb of
-              Left  err  -> Left err
-              Right _rDb -> Right Runtime { .. }
-
--- | Create a runtime from a given state.
-boot' :: MonadIO m => Data.HaskDb -> FilePath -> m Runtime
-boot' db logsPath = do
-  let loggingConf = Command.mkLoggingConf logsPath
-      _rConf      = Command.defaultConf { Command._confLogging = loggingConf }
-  _rDb      <- liftIO . STM.atomically $ Core.instantiateStmDb db
-  _rLoggers <- ML.makeDefaultLoggersWithConf loggingConf
-  let _rThreads = NoThreads
-  pure $ Runtime { .. }
-
-{- | Instantiate the db.
-
-1. The state is either the empty db, or if a _confDbFile file is specified, is
-read from the file.
-
-2. Whenever the application exits, the state is written to disk, if a
-_confDbFile is specified.
--}
-instantiateDb
-  :: forall m
-   . MonadIO m
-  => Command.Conf
-  -> m (Either Errs.RuntimeErr Core.StmDb)
-instantiateDb Command.Conf {..} = readDbSafe _confDbFile
-
-readDb
-  :: forall m
-   . MonadIO m
-  => Maybe FilePath
-  -> m (Either Errs.RuntimeErr Core.StmDb)
-readDb mpath = case mpath of
-  Just fpath -> do
-    -- We may want to read the file only when the file exists.
-    exists <- liftIO $ doesFileExist fpath
-    if exists then fromFile fpath else useEmpty
-  Nothing -> useEmpty
- where
-  fromFile fpath = liftIO (try @SomeException $ T.readFile fpath) >>= \case
-    Left err ->
-      putStrLn @Text ("Unable to read db file: " <> maybe "" T.pack mpath)
-        $> Left (Errs.RuntimeException err)
-    Right fdata ->
-           -- We may want to deserialise the data only when the data is non-empty.
-                   if T.null fdata
-      then useEmpty
-      else
-        Data.deserialiseDbStrict (TE.encodeUtf8 fdata)
-          & either (pure . Left . Errs.knownErr) useState
-  useEmpty = Right <$> liftIO (STM.atomically Core.instantiateEmptyStmDb)
-  useState = fmap Right . liftIO . STM.atomically . Core.instantiateStmDb
-
--- | A safer version of readDb: this fails if the file doesn't exist or is
--- empty. This helps in catching early mistake, e.g. from user specifying the
--- wrong file name on the command-line.
-readDbSafe
-  :: forall m
-   . MonadIO m
-  => Maybe FilePath
-  -> m (Either Errs.RuntimeErr Core.StmDb)
-readDbSafe mpath = case mpath of
-  Just fpath -> do
-    -- We may want to read the file only when the file exists.
-    exists <- liftIO $ doesFileExist fpath
-    if exists
-      then fromFile fpath
-      else pure . Left . Errs.knownErr $ RErr.FileDoesntExistErr fpath
-  Nothing -> useEmpty
- where
-  fromFile fpath = do
-    fdata <- liftIO (T.readFile fpath)
-    Data.deserialiseDbStrict (TE.encodeUtf8 fdata)
-      & either (pure . Left . Errs.knownErr) useState
-  useEmpty = Right <$> liftIO (STM.atomically Core.instantiateEmptyStmDb)
-  useState = fmap Right . liftIO . STM.atomically . Core.instantiateStmDb
